Updated squareroot

+++ MyMath.java
@@ -8,9 +8,15 @@ class MyMath {
 
   
 
-  public static double squareroot(double inpt) {
+  public static int squareroot(int inpt) {
 
-    return 0.0;
+    // OPTIONAL!!!!! Just for fun!!!
+    // Step 1: Just use a method from the Math class to make sure it works
+    // Step 2:
+    // use a "for" loop to start at 1 and go up and up to see if the square of your number == input
+    // if you get a square that's higher than input, just say "the input is not a perfect square!"
+
+    return 0;
 
   }
